site/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
django-rosetta==0.10.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
environs==10.3.0
idna==3.6
marshmallow==3.20.2
packaging==23.2
pillow==10.2.0
polib==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
